requirements.txt


fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
Cmake
opencv-python
pandas
pyyaml
psutil
tqdm
matplotlib
seaborn
gitpython==3.1.30
setuptools==65.5.1
